jd_red_flag_analyzer/requirements.txt
streamlit>=1.30
pydantic>=2.5
google-genai>=1.0
tavily-python>=0.5
python-dotenv>=1.0
plotly>=5.18
pandas>=2.0
pytest>=8.0
